modify the implementation of rest api

+++ machineRestapi/src/main/java/com/cn/ibm/miniist/tingtingtang/rest/RestAPI.java
@@ -33,42 +33,11 @@ public class RestAPI {
 	}
 
 	/*
-	 * This put is used to create a new machine info to the database
+	 * This POST is used to create a new machine info to the database
 	 */
 	
 	@POST
-	@Path("/test/{mName}")
-	@Consumes(MediaType.APPLICATION_FORM_URLENCODED)
-	public Response testDBConnection(
-			@PathParam("mName") String mName)
-	{
-		MachineInfo machineInfo = new MachineInfo();
-		MachineInfoService machineInfoService = new MachineInfoService();
-		try 
-		{
-			if (mName != null) 
-			{
-				machineInfo = machineInfoService.findMachine(mName, "mlist");
-
-				Gson gson = new Gson();
-				String json = gson.toJson(machineInfo);
-				return Response.status(201).entity(json).build();
-
-			} 
-			else
-			{
-				return Response.status(Status.BAD_REQUEST).build();
-			}
-		}
-		catch (Exception e) 
-		{
-			return Response.status(Status.BAD_REQUEST).build();
-		}
-		
-	}
-	
-	@PUT
-	@Path("/create")
+	@Path("/")
 	@Consumes(MediaType.APPLICATION_FORM_URLENCODED)
 	public Response createMachineInfo(
 			@FormParam("userName") String userName,
@@ -77,11 +46,12 @@ public class RestAPI {
 			@FormParam("ipAddress") String ipAddress,
 			@FormParam("description") String description)
 	{
-		MachineInfo machineInfo = new MachineInfo();
-		MachineInfoService machineInfoService = new MachineInfoService();
-		String createTime;
-		SimpleDateFormat dataFormat = new SimpleDateFormat("yyyy-MM-dd HH:MM:SS");
-		try{
+		if(userName != null & mName != null)
+		{
+			MachineInfo machineInfo = new MachineInfo();
+			MachineInfoService machineInfoService = new MachineInfoService();
+			String createTime;
+			SimpleDateFormat dataFormat = new SimpleDateFormat("yyyy-MM-dd HH:MM:SS");
 			createTime = dataFormat.format(new Timestamp(new java.util.Date().getTime()));
 			machineInfo.setUserName(userName);
 			machineInfo.setMachineName(mName);
@@ -90,30 +60,26 @@ public class RestAPI {
 			machineInfo.setDes(description);
 			machineInfo.setCreateTime(createTime);
 			machineInfo.setUpdateTime(createTime);
-			
-			if(machineInfoService.createNewMachine(machineInfo, "mlist"))
+
+			if (machineInfoService.createNewMachine(machineInfo, "mlist")) 
 			{
-				return Response.status(201).entity("Succeed!").build();
+				return Response.status(201).entity("New machine has been created!").build();
 			}
-			else
-			{
-				return Response.status(Status.BAD_REQUEST).build();
-			}
+			else 
+				return Response.status(Status.BAD_REQUEST).entity("The machine has already been in DB!").build();
 		}
-		catch (Exception e) 
-		{
-			return Response.status(Status.BAD_REQUEST).build();
-		}
+		else
+			return Response.status(Status.BAD_REQUEST).entity("The userName or mName is null!").build();
 	}
 	
 	/*
-	 * This post is used to find the machine info according to the unique machine name
+	 * This GET is used to find the machine info according to the unique machine name
 	 */
-	@POST
-	@Path("/find")
+	@GET
+	@Path("/machine/{mName}")
 	@Consumes(MediaType.APPLICATION_FORM_URLENCODED)
 	public Response findMachineInfo(
-			@FormParam("mName") String mName
+			@PathParam("mName") String mName
 			) 
 	{
 		MachineInfo machineInfo = new MachineInfo();
@@ -122,50 +88,55 @@ public class RestAPI {
 			if (mName != null) 
 			{
 				machineInfo = machineInfoService.findMachine(mName, "mlist");
-
-				Gson gson = new Gson();
-				String json = gson.toJson(machineInfo);
 				
-				return Response.status(201).entity(json).build();
+				if(machineInfo != null)
+				{
+					Gson gson = new Gson();
+					String json = gson.toJson(machineInfo);
+					
+					return Response.status(201).entity(json).build();
+				}
+				else return Response.status(Status.BAD_REQUEST).entity("The mName does not exist in DB!").build();
 			} 
 			else
-			{
-				return Response.status(Status.BAD_REQUEST).build();
-			}
+				return Response.status(Status.BAD_REQUEST).entity("The mName is null").build();
 
 		} 
 		catch (Exception e) 
 		{
-			return Response.status(Status.BAD_REQUEST).build();
+			return Response.status(Status.BAD_REQUEST).entity("exception").build();
 		}
 	}
 	
 	/*
-	 * This post is used to find the machine list under specific user
+	 * This GET is used to find the machine list under specific user
 	 */
-	@POST
-	@Path("/find/mlist")
+	@GET
+	@Path("/mlist/{userName}")
 	@Consumes(MediaType.APPLICATION_FORM_URLENCODED)
 	public Response findMachineList(
-			@FormParam("userName") String userName
+			@PathParam("userName") String userName
 			) 
-	{
+	{		
 		List<MachineInfo> machineList = new ArrayList<MachineInfo>();
 		MachineInfoService machineInfoService = new MachineInfoService();
 		try {
 			if (userName != null) 
 			{
 				machineList = machineInfoService.findMachineList(userName, "mlist");
-
-				Gson gson = new Gson();
-				String json = gson.toJson(machineList.toArray());
 				
-				return Response.status(201).entity(json).build();
+				if (machineList != null)
+				{
+					Gson gson = new Gson();
+					String json = gson.toJson(machineList.toArray());
+					
+					return Response.status(201).entity(json).build();
+				}
+				else
+					return Response.status(Status.BAD_REQUEST).entity("There is no machine of the user!").build();
 			} 
 			else
-			{
-				return Response.status(Status.BAD_REQUEST).build();
-			}
+				return Response.status(Status.BAD_REQUEST).entity("The userName is null!").build();
 
 		} 
 		catch (Exception e) 
@@ -174,12 +145,11 @@ public class RestAPI {
 		}
 	}
 	
-	
 	/*
-	 * Update machine info from the database
+	 * This PUT is used to update machine info from the database
 	 */
-	@POST
-	@Path("/update")
+	@PUT
+	@Path("/")
 	@Consumes(MediaType.APPLICATION_FORM_URLENCODED)
 	public Response findMachineInfo(
 			@FormParam("userName") String userName,
@@ -189,57 +159,70 @@ public class RestAPI {
 			@FormParam("description") String description
 			) 
 	{
-		MachineInfo machineInfo = new MachineInfo();
-		MachineInfoService machineInfoService = new MachineInfoService();
-		String updateTime;
-		SimpleDateFormat dataFormat = new SimpleDateFormat("yyyy-MM-dd HH:MM:SS");
-		try{
-			updateTime = dataFormat.format(new Timestamp(new java.util.Date().getTime()));
-			machineInfo.setUserName(userName);;
-			machineInfo.setMachineName(mName);
-			machineInfo.setHostName(hostName);
-			machineInfo.setIpAdd(ipAddress);
-			machineInfo.setDes(description);
-			machineInfo.setUpdateTime(updateTime);
-			
-			if(machineInfoService.updateMachine(machineInfo, "mlist"))
+		if(userName != null && mName != null)
+		{
+			MachineInfo machineInfo = new MachineInfo();
+			MachineInfoService machineInfoService = new MachineInfoService();
+			String updateTime;
+			SimpleDateFormat dataFormat = new SimpleDateFormat("yyyy-MM-dd HH:MM:SS");
+			try
 			{
-				return Response.status(201).entity("Succeed!").build();
+				updateTime = dataFormat.format(new Timestamp(new java.util.Date().getTime()));
+				machineInfo.setUserName(userName);;
+				machineInfo.setMachineName(mName);
+				machineInfo.setHostName(hostName);
+				machineInfo.setIpAdd(ipAddress);
+				machineInfo.setDes(description);
+				machineInfo.setUpdateTime(updateTime);
+				
+				if(machineInfoService.updateMachine(machineInfo, "mlist"))
+				{
+					return Response.status(201).entity("DB has been updated!").build();
+				}
+				else
+				{
+					return Response.status(Status.BAD_REQUEST).entity("The mName or the userName does not exist in DB!").build();
+				}
 			}
-			else
+			catch (Exception e) 
 			{
-				return Response.status(Status.BAD_REQUEST).build();
+				return Response.status(Status.BAD_REQUEST).entity("exception!").build();
 			}
 		}
-		catch (Exception e) 
+		else
 		{
-			return Response.status(Status.BAD_REQUEST).build();
+			return Response.status(Status.BAD_REQUEST).entity("The userName or the mName is null!").build();
 		}
 	}
 	/*
-	 * Delete the machine from the database
+	 * This DELETE is used to delete the machine from the database
 	 */
 	@DELETE
-	@Path("/delete/{mName}")
+	@Path("/{mName}")
 	@Consumes(MediaType.APPLICATION_FORM_URLENCODED)
 	public Response deleteMachineInfo(
 			@PathParam("mName") String mName)
 	{
-		MachineInfoService machineInfoService = new MachineInfoService();
-		try
+		if(mName != null)
 		{
-			if(machineInfoService.deleteMachine(mName, "mlist"))
+			MachineInfoService machineInfoService = new MachineInfoService();
+			try
 			{
-				return Response.status(200).entity("delete" + mName + "successfully !").build();
+				if(machineInfoService.deleteMachine(mName, "mlist"))
+				{
+					return Response.status(200).entity("The machine has been deleted!").build();
+				}
+				else
+				{
+					return Response.status(Status.BAD_REQUEST).entity("The machine does not exist in DB!").build();
+				}	
 			}
-			else
+			catch (Exception e)
 			{
-				return Response.status(Status.BAD_REQUEST).entity(mName + "is not exist in the database!").build();
-			}	
+				return Response.status(Status.BAD_REQUEST).entity("exception!").build();
+			}
 		}
-		catch (Exception e)
-		{
-			return Response.status(Status.BAD_REQUEST).build();
-		}
+		else
+			return Response.status(Status.BAD_REQUEST).entity("The mName is null!").build();
 	}
 }
